+++ src/Routes.tsx
@@ -1,9 +1,19 @@
 import { DebugProvider } from "./DebugContext";
 import { Dev } from "./components/Dev/Dev";
 import { Yjs } from "./components/Dev/Yjs";
-import { Layout } from "./components/Layout";
 import Editor from "@/components/Editor/Editor";
 import { Routes as RouterRoutes, Route } from "react-router-dom";
+import { Outlet } from "react-router-dom";
+import { TopNavBar } from "./components/TopNavBar";
+
+function Layout() {
+  return (
+    <div className="container">
+      <TopNavBar />
+      <Outlet />
+    </div>
+  );
+}
 
 export function Routes() {
   return (
