+++ src/index.tsx
@@ -1,5 +1,5 @@
+/// <reference types="vite/client" />
 import App from "./App";
-import React from "react";
 import ReactDOM from "react-dom/client";
 
 //clear console on hot reload
@@ -7,6 +7,6 @@ if (import.meta.hot) {
   import.meta.hot.on("vite:beforeUpdate", () => console.clear());
 }
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById("root")!);
 
 root.render(<App />);
